\name{MuFiCokriging-package}
\alias{MuFiCokriging-package}
\alias{MuFiCokriging}
\docType{package}
\title{
Multi-Fidelity Cokriging methods for computer  experiments
}
\description{
Create multi-fidelity cokriging models using data from codes with multiple levels of fidelity.
}
\details{
\tabular{ll}{
Package: \tab MuFiCokriging\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-21\cr
License: \tab GPL-3 \cr
}

\tabular{l}{
Important functions :\cr \cr

MuFicokm : \cr

 Creation of a multi-fidelity cokriging model with unknown or known parameters. \cr \cr

predict.MuFicokm :\cr

Prediction of a multi-fidelity cokriging model at new points (Simple and Universal cokriging). \cr
}
}
\author{
Loic Le Gratiet

Maintainer: <loic.legratiet@gmail.com>  - 
Loic Le Gratiet

Affiliations: \cr
Universite Paris VII Denis-Diderot 75205 Paris Cedex 13 \cr
CEA, DAM, DIF, F-91297 Arpajon, France
}
\references{

KENNEDY, M.C. & O'HAGAN, A. (2000), Predicting the output from a complex computer code when fast approximations are available. \emph{Biometrika} 87, 1-13

FORRESTER, A.I.J, SOBESTER A. & KEAN, A.J. (2007), Multi-Fidelity optimization via surrogate modelling. \emph{Proc. R. Soc. A} 463, 3251-3269.

LE GRATIET, L. & GARNIER, J. (2012), Recursive co-kriging model for Design of Computer Experiments with multiple levels of fidelity, \emph{arXiv:1210.0686}

LE GRATIET, L. (2012), Bayesian analysis of hierarchical multi-fidelity codes, \emph{arXiv:1112.5389}

}
